fix: split validation module, fix auth and dry-run bugs
- extract validateSyncConfig into sync/validate.ts (run.ts under 500 lines)
- resolveSyncAuth returns all missing env var names, not just first
- SyncSourceError pre-flight now uses src.type, not empty string
- dry-run handled at orchestrator level; runSource no longer casts it
- scope validated for github, google_workspace, gerrit adapters
