+++ packages/engram-core/src/sync/run.ts
@@ -17,7 +17,7 @@ import {
   loadExecutablePlugin,
   loadJsModulePlugin,
 } from "../plugins/index.js";
-import { SyncConfigValidationError, SyncSourceError } from "./errors.js";
+import { SyncSourceError } from "./errors.js";
 import type {
   RunSyncOpts,
   SourceResult,
@@ -28,254 +28,13 @@ import type {
 } from "./types.js";
 import { resolveSyncAuth } from "./types.js";
 
+export { validateSyncConfig } from "./validate.js";
+
 // ---------------------------------------------------------------------------
 // Built-in adapter types
 // ---------------------------------------------------------------------------
 
 const BUILT_IN_TYPES = ["git", "source", "github", "google_workspace"] as const;
-
-// ---------------------------------------------------------------------------
-// Config validation
-// ---------------------------------------------------------------------------
-
-const ALLOWED_AUTH_KINDS = [
-  "none",
-  "bearer",
-  "basic",
-  "service_account",
-  "oauth2",
-] as const;
-
-const ALLOWED_SOURCE_FIELDS = new Set([
-  "name",
-  "type",
-  "scope",
-  "path",
-  "root",
-  "auth",
-]);
-
-const ALLOWED_AUTH_FIELDS: Record<string, Set<string>> = {
-  none: new Set(["kind"]),
-  bearer: new Set(["kind", "tokenEnv"]),
-  basic: new Set(["kind", "usernameEnv", "secretEnv"]),
-  service_account: new Set(["kind", "keyJsonEnv"]),
-  oauth2: new Set(["kind", "tokenEnv", "scopesEnv"]),
-};
-
-/**
- * Validate a raw JSON object against the SyncConfig schema.
- * Collects ALL errors before throwing, so the user can fix everything at once.
- *
- * Rules (fail-closed):
- * - `version: 1` required
- * - Every source: `name` (unique), `type` — no unknown fields
- * - Auth uses `SyncAuthConfig` union — no unknown fields per kind
- */
-export function validateSyncConfig(raw: unknown): SyncConfig {
-  if (typeof raw !== "object" || raw === null || Array.isArray(raw)) {
-    throw new SyncConfigValidationError([
-      { field: "(root)", reason: "must be a JSON object" },
-    ]);
-  }
-
-  const obj = raw as Record<string, unknown>;
-  const failures: Array<{ field: string; reason: string }> = [];
-
-  // Check for unknown top-level fields
-  const ALLOWED_TOP_FIELDS = new Set(["version", "sources"]);
-  for (const key of Object.keys(obj)) {
-    if (!ALLOWED_TOP_FIELDS.has(key)) {
-      failures.push({
-        field: key,
-        reason: `unknown top-level field (allowed: ${Array.from(ALLOWED_TOP_FIELDS).join(", ")})`,
-      });
-    }
-  }
-
-  // Validate version
-  if (!("version" in obj)) {
-    failures.push({ field: "version", reason: "required field missing" });
-  } else if (obj.version !== 1) {
-    failures.push({
-      field: "version",
-      reason: `must be 1 (got ${JSON.stringify(obj.version)}). If you have a newer config format, upgrade engram to the matching version.`,
-    });
-  }
-
-  // Validate sources
-  if (!("sources" in obj)) {
-    failures.push({ field: "sources", reason: "required field missing" });
-  } else if (!Array.isArray(obj.sources)) {
-    failures.push({ field: "sources", reason: "must be an array" });
-  } else {
-    const seenNames = new Set<string>();
-
-    for (let i = 0; i < obj.sources.length; i++) {
-      const src = obj.sources[i];
-      const prefix = `sources[${i}]`;
-
-      if (typeof src !== "object" || src === null || Array.isArray(src)) {
-        failures.push({ field: prefix, reason: "must be an object" });
-        continue;
-      }
-
-      const srcObj = src as Record<string, unknown>;
-
-      // Check for unknown source fields
-      for (const key of Object.keys(srcObj)) {
-        if (!ALLOWED_SOURCE_FIELDS.has(key)) {
-          failures.push({
-            field: `${prefix}.${key}`,
-            reason: `unknown field (allowed: ${Array.from(ALLOWED_SOURCE_FIELDS).join(", ")})`,
-          });
-        }
-      }
-
-      // name
-      if (!("name" in srcObj)) {
-        failures.push({ field: `${prefix}.name`, reason: "required" });
-      } else if (typeof srcObj.name !== "string" || !srcObj.name) {
-        failures.push({
-          field: `${prefix}.name`,
-          reason: "must be a non-empty string",
-        });
-      } else if (seenNames.has(srcObj.name)) {
-        failures.push({
-          field: `${prefix}.name`,
-          reason: `duplicate source name '${srcObj.name}' — names must be unique`,
-        });
-      } else {
-        seenNames.add(srcObj.name as string);
-      }
-
-      // type
-      if (!("type" in srcObj)) {
-        failures.push({ field: `${prefix}.type`, reason: "required" });
-      } else if (typeof srcObj.type !== "string" || !srcObj.type) {
-        failures.push({
-          field: `${prefix}.type`,
-          reason: "must be a non-empty string",
-        });
-      }
-
-      // scope (optional string)
-      if ("scope" in srcObj && typeof srcObj.scope !== "string") {
-        failures.push({ field: `${prefix}.scope`, reason: "must be a string" });
-      }
-
-      // path (optional string)
-      if ("path" in srcObj && typeof srcObj.path !== "string") {
-        failures.push({ field: `${prefix}.path`, reason: "must be a string" });
-      }
-
-      // root (optional string)
-      if ("root" in srcObj && typeof srcObj.root !== "string") {
-        failures.push({ field: `${prefix}.root`, reason: "must be a string" });
-      }
-
-      // auth (optional)
-      if ("auth" in srcObj) {
-        const authFailures = validateAuthConfig(srcObj.auth, `${prefix}.auth`);
-        failures.push(...authFailures);
-      }
-    }
-  }
-
-  if (failures.length > 0) {
-    throw new SyncConfigValidationError(failures);
-  }
-
-  return obj as unknown as SyncConfig;
-}
-
-function validateAuthConfig(
-  auth: unknown,
-  fieldPrefix: string,
-): Array<{ field: string; reason: string }> {
-  const failures: Array<{ field: string; reason: string }> = [];
-
-  if (typeof auth !== "object" || auth === null || Array.isArray(auth)) {
-    failures.push({ field: fieldPrefix, reason: "must be an object" });
-    return failures;
-  }
-
-  const authObj = auth as Record<string, unknown>;
-
-  if (!("kind" in authObj)) {
-    failures.push({ field: `${fieldPrefix}.kind`, reason: "required" });
-    return failures;
-  }
-
-  const kind = authObj.kind;
-  if (
-    !ALLOWED_AUTH_KINDS.includes(kind as (typeof ALLOWED_AUTH_KINDS)[number])
-  ) {
-    failures.push({
-      field: `${fieldPrefix}.kind`,
-      reason: `must be one of [${ALLOWED_AUTH_KINDS.join(", ")}], got ${JSON.stringify(kind)}`,
-    });
-    return failures;
-  }
-
-  const kindStr = kind as string;
-  const allowedFields = ALLOWED_AUTH_FIELDS[kindStr];
-  if (allowedFields) {
-    for (const key of Object.keys(authObj)) {
-      if (!allowedFields.has(key)) {
-        failures.push({
-          field: `${fieldPrefix}.${key}`,
-          reason: `unknown field for auth kind '${kindStr}' (allowed: ${Array.from(allowedFields).join(", ")})`,
-        });
-      }
-    }
-  }
-
-  // Kind-specific required fields
-  switch (kindStr) {
-    case "bearer":
-      if (typeof authObj.tokenEnv !== "string" || !authObj.tokenEnv) {
-        failures.push({
-          field: `${fieldPrefix}.tokenEnv`,
-          reason:
-            "required for bearer auth — name of the env var holding the token",
-        });
-      }
-      break;
-    case "basic":
-      if (typeof authObj.usernameEnv !== "string" || !authObj.usernameEnv) {
-        failures.push({
-          field: `${fieldPrefix}.usernameEnv`,
-          reason: "required for basic auth",
-        });
-      }
-      if (typeof authObj.secretEnv !== "string" || !authObj.secretEnv) {
-        failures.push({
-          field: `${fieldPrefix}.secretEnv`,
-          reason: "required for basic auth",
-        });
-      }
-      break;
-    case "service_account":
-      if (typeof authObj.keyJsonEnv !== "string" || !authObj.keyJsonEnv) {
-        failures.push({
-          field: `${fieldPrefix}.keyJsonEnv`,
-          reason: "required for service_account auth",
-        });
-      }
-      break;
-    case "oauth2":
-      if (typeof authObj.tokenEnv !== "string" || !authObj.tokenEnv) {
-        failures.push({
-          field: `${fieldPrefix}.tokenEnv`,
-          reason: "required for oauth2 auth",
-        });
-      }
-      break;
-  }
-
-  return failures;
-}
 
 // ---------------------------------------------------------------------------
 // Auth env var pre-flight check
@@ -292,9 +51,9 @@ function checkAuthEnvVars(
 
   for (const src of sources) {
     if (!src.auth) continue;
-    const result = resolveSyncAuth(src.auth);
-    if ("missing" in result) {
-      missing.push({ sourceName: src.name, envVar: result.missing });
+    const missingVars = resolveSyncAuth(src.auth);
+    for (const envVar of missingVars) {
+      missing.push({ sourceName: src.name, envVar });
     }
   }
 
@@ -373,7 +132,6 @@ async function runSource(
   graph: EngramGraph,
   src: SyncSource,
   cwd: string,
-  dryRun: boolean,
 ): Promise<SourceResult> {
   const startMs = Date.now();
   const type = src.type;
@@ -381,9 +139,7 @@ async function runSource(
   try {
     if (type === "git") {
       const repoPath = path.resolve(cwd, src.path ?? src.root ?? ".");
-      const result = await ingestGitRepo(graph, repoPath, {
-        dryRun,
-      } as Parameters<typeof ingestGitRepo>[2]);
+      const result = await ingestGitRepo(graph, repoPath);
       return {
         name: src.name,
         type,
@@ -397,7 +153,7 @@ async function runSource(
 
     if (type === "source") {
       const root = path.resolve(cwd, src.root ?? src.path ?? ".");
-      const result = await ingestSource(graph, { root, dryRun });
+      const result = await ingestSource(graph, { root });
       return {
         name: src.name,
         type,
@@ -415,13 +171,15 @@ async function runSource(
     // Resolve auth credential
     let auth: AuthCredential = { kind: "none" };
     if (src.auth) {
-      const resolved = resolveSyncAuth(src.auth as SyncAuthConfig);
-      if ("missing" in resolved) {
+      const missing = resolveSyncAuth(src.auth as SyncAuthConfig);
+      if (missing.length > 0) {
         throw new Error(
-          `Missing required env var '${resolved.missing}' for source '${src.name}'`,
+          `Missing required env vars for source '${src.name}': ${missing.join(", ")}`,
         );
       }
-      auth = resolved.credential;
+      // Re-resolve to get the credential (all vars confirmed present)
+      const resolved = resolveSyncAuthCredential(src.auth as SyncAuthConfig);
+      auth = resolved;
     }
 
     // Validate scope
@@ -432,7 +190,6 @@ async function runSource(
     const result = await adapter.enrich(graph, {
       auth,
       scope: src.scope,
-      dryRun,
     });
 
     return {
@@ -492,9 +249,14 @@ export async function runSync(
     const lines = missingEnvVars
       .map((m) => `  source '${m.sourceName}': missing env var '${m.envVar}'`)
       .join("\n");
+    // B3: use the actual source type from the first failing source
+    const firstMissing = missingEnvVars[0];
+    const firstSrc = sourcesToRun.find(
+      (s) => s.name === firstMissing.sourceName,
+    );
     throw new SyncSourceError(
-      missingEnvVars[0].sourceName,
-      "",
+      firstMissing.sourceName,
+      firstSrc?.type ?? "",
       new Error(
         `Missing required environment variables:\n${lines}\n` +
           `Set these env vars before running sync.`,
@@ -518,6 +280,7 @@ export async function runSync(
 
     opts.onSourceStart?.(src.name, src.type);
 
+    // B4: dry-run is handled entirely at orchestrator level — skip adapter calls
     const result = dryRun
       ? ({
           name: src.name,
@@ -528,7 +291,7 @@ export async function runSync(
           edgesCreated: 0,
           elapsedMs: 0,
         } satisfies SourceResult)
-      : await runSource(graph, src, cwd, dryRun);
+      : await runSource(graph, src, cwd);
 
     sourceResults.push(result);
     opts.onSourceEnd?.(result);
@@ -567,4 +330,42 @@ export async function runSync(
     status: overallStatus,
     elapsedMs: Date.now() - totalStart,
   };
+}
+
+// ---------------------------------------------------------------------------
+// Internal: resolve credential after all vars confirmed present
+// ---------------------------------------------------------------------------
+
+function resolveSyncAuthCredential(authConfig: SyncAuthConfig): AuthCredential {
+  switch (authConfig.kind) {
+    case "none":
+      return { kind: "none" };
+    case "bearer":
+      return {
+        kind: "bearer",
+        token: process.env[authConfig.tokenEnv] ?? "",
+      };
+    case "basic":
+      return {
+        kind: "basic",
+        username: process.env[authConfig.usernameEnv] ?? "",
+        secret: process.env[authConfig.secretEnv] ?? "",
+      };
+    case "service_account":
+      return {
+        kind: "service_account",
+        keyJson: process.env[authConfig.keyJsonEnv] ?? "",
+      };
+    case "oauth2": {
+      const rawScopes = authConfig.scopesEnv
+        ? process.env[authConfig.scopesEnv]
+        : undefined;
+      const scopes = rawScopes ? rawScopes.split(",").map((s) => s.trim()) : [];
+      return {
+        kind: "oauth2",
+        token: process.env[authConfig.tokenEnv] ?? "",
+        scopes,
+      };
+    }
+  }
 }
